refactor: history helper file
