+++ app/src/main/kotlin/xyz/lbres/trickcalculator/ui/history/HistoryViewModel.kt
@@ -1,13 +1,7 @@
 package xyz.lbres.trickcalculator.ui.history
 
 import androidx.lifecycle.ViewModel
-import xyz.lbres.exactnumbers.exactfraction.ExactFraction
-import xyz.lbres.kotlinutils.general.simpleIf
-import xyz.lbres.kotlinutils.list.StringList
-import xyz.lbres.kotlinutils.random.ext.nextBoolean
-import xyz.lbres.trickcalculator.SharedValues.random
 import xyz.lbres.trickcalculator.utils.History
-import xyz.lbres.trickcalculator.utils.seededShuffled
 
 /**
  * Information about computation history
@@ -21,7 +15,7 @@ class HistoryViewModel : ViewModel() {
         get() = _randomness
         set(value) {
             _randomness = value
-            updateRandomHistory()
+            randomizedHistory = generateRandomHistory(history, value)
         }
 
     /**
@@ -35,57 +29,6 @@ class HistoryViewModel : ViewModel() {
      */
     var randomizedHistory: History? = null
         private set
-
-    /**
-     * Update randomized history based on current randomness
-     */
-    private fun updateRandomHistory() {
-        randomizedHistory = when (randomness) {
-            0 -> history // no randomness
-            1 -> history.seededShuffled() // shuffled order
-            2 -> shuffleHistoryValues() // shuffled values
-            3 -> generateRandomHistory() ?: emptyList() // random generation
-            else -> history
-        }
-    }
-
-    /**
-     * Shuffle history computations and results/errors.
-     * Returns a list that contains all computations and results/errors, but not necessarily in their original pari.
-     *
-     * @return [History]: history where computations and values have been shuffled
-     */
-    private fun shuffleHistoryValues(): History {
-        val computations: List<StringList> = history.map { it.computation }.seededShuffled()
-        val values: List<Pair<ExactFraction?, String?>> = history.map { Pair(it.result, it.error) }.seededShuffled()
-
-        val shuffledHistory = computations.mapIndexed { index, comp ->
-            val valuePair = values[index]
-            if (valuePair.first != null) {
-                HistoryItem(comp, valuePair.first!!)
-            } else {
-                HistoryItem(comp, valuePair.second!!)
-            }
-        }
-
-        return shuffledHistory
-    }
-
-    /**
-     * Generate randomized history items based on the length of the input.
-     * Returns null randomly or if history is empty, to indicate an "error" in retrieving history.
-     *
-     * @return [History]: possibly null history of generated computations, with same length as real history (if not null)
-     */
-    private fun generateRandomHistory(): History? {
-        val probabilityError = simpleIf(history.isEmpty(), 1f, 0.2f)
-
-        if (random.nextBoolean(probabilityError)) {
-            return null
-        }
-
-        return List(history.size) { generateRandomHistoryItem() }
-    }
 
     /**
      * Add new item to history
