+++ app/src/main/kotlin/xyz/lbres/trickcalculator/ui/history/generateRandomHistoryItem.kt
@@ -146,6 +146,7 @@ private fun generateErrorMessage(): String {
         "Number overflow exception"
     )
 
+    // TODO use seeded values
     val message = messages.random()
 
     if (message == "Number overflow exception") {
