+++ app/src/test/kotlin/xyz/lbres/trickcalculator/ui/history/GenerateRandomHistoryItemTest.kt
@@ -1,0 +1,10 @@
+package xyz.lbres.trickcalculator.ui.history
+
+import kotlin.test.Test
+
+class GenerateRandomHistoryItemTest {
+    @Test
+    fun testGenerateRandomHistoryItem() {
+        // TODO
+    }
+}
